Rewrite _get_braced_expr for text to not depend on compile()
Prior to this commit, the _get_braced_expr implementation for text
templates depended on compile() to raise a syntax error with the
"offset" attribute set at the ending brace, even if there were further
syntax errors in the expression.

This is true of the implementation of compile for CPython 3.10b1 and
earlier.  However, in 3.10b2, the implementation changed, and now the
offset of the syntax error is in the wrong position:

Python 3.9 and earlier:

  >>> compile('caller(i)}."', '', 'eval')
  Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
    File "", line 1
      caller(i)}."
               ^
  SyntaxError: unmatched '}'

Python 3.10b2 (and presumably future releases of Python):

  >>> compile('caller(i)}."', '', 'eval')
  Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
    File "", line 1
      caller(i)}."
                 ^
  SyntaxError: unterminated string literal (detected at line 1)

This seems like an internal detail of how compile() works, and
probably best to not depend on it!

I re-wrote _get_braced_expr here so that it uses the tokenize module
instead, and counts brace tokens (expecting matching tokens).  This
seems like a considerably safer and more portable way to implement the
functionality.

FIXES=#49