Switch the test runner to Pytest
Nose has compatibility issues with Python 3.10, and isn't being
updated.  The pytest runner seems to work with the existing tests, so
switch to it.

Additionally, "python setup.py test" is being deprecated too, so get
rid of that.  In the future, we may want to take a look at tox.

Resolves #54.