Fix lnotab bugs in Python 3.6+
In Python 3.6, the lnotab can contain negative line offsets, as this
is now a signed byte instead of unsigned.

Python 3.4+ provides a function dis.findlinestarts, which does the
exact same thing as lnotab_numbers, but understands the
signed/unsigned distinguishment in the bytecode.  This requires we
adjust the lnotab implementation to generate a new lnotab with signed
bytes as well.

Finally, to plug up all the ERROR LOOKING UP LINE #XXX messages, fix a
bug where we don't look a newline characters in the generated python
lines, that way the numbers get mapped correctly.